+++ pkg/infra/usagestats/statscollector/service_test.go
@@ -177,6 +177,7 @@ func TestCollectingUsageStats(t *testing.T) {
 
 	assert.EqualValues(t, 11, metrics["stats.data_keys.count"])
 	assert.EqualValues(t, 3, metrics["stats.active_data_keys.count"])
+	assert.EqualValues(t, 5, metrics["stats.public_dashboards.count"])
 
 	assert.InDelta(t, int64(65), metrics["stats.uptime"], 6)
 }
@@ -351,6 +352,7 @@ func mockSystemStats(sqlStore *mockstore.SQLStoreMock) {
 		APIKeys:                   2,
 		DataKeys:                  11,
 		ActiveDataKeys:            3,
+		PublicDashboards:          5,
 	}
 }